+++ src/app/projects/[projectId]/columns.tsx
@@ -20,11 +20,11 @@ export const columns: ColumnDef<Dependency>[] = [
     header: 'Tool',
   },
   {
-    accessorKey: 'value',
-    header: 'Version',
-  },
-  {
     accessorKey: 'modified_at',
     header: 'Updated',
+  },
+  {
+    accessorKey: 'value',
+    header: 'Version',
   },
 ]
