+++ src/utils/utility-functions.ts
@@ -1,0 +1,22 @@
+export const extractCycle = (version: string): string => {
+  const parts = version.split('.')
+  return parts.slice(0, 2).join('.')
+}
+
+export const formatDate = (dateString: string): string => {
+  const now = new Date()
+  const date = new Date(dateString)
+  const diffInSeconds = Math.floor((now.getTime() - date.getTime()) / 1000)
+
+  if (diffInSeconds < 60) return 'just now'
+  if (diffInSeconds < 3600)
+    return `${Math.floor(diffInSeconds / 60)} minutes ago`
+  if (diffInSeconds < 86400)
+    return `${Math.floor(diffInSeconds / 3600)} hours ago`
+  if (diffInSeconds < 2592000)
+    return `${Math.floor(diffInSeconds / 86400)} days ago`
+  if (diffInSeconds < 31536000)
+    return `${Math.floor(diffInSeconds / 2592000)} months ago`
+
+  return `${Math.floor(diffInSeconds / 31536000)} years ago`
+}
